Skyline: Fix a few warnings in the build.
Former-commit-id: 78ad0fa1f80ff58840178e6941d94ede255552bc

+++ pwiz_tools/Shared/Crawdad/filters/GaussSmoother.cpp
@@ -84,7 +84,7 @@ void GaussSmoother::trim_weights_by_frac_max(float frac ) {
             break;
         }
     }
-    for ( size_t i = weights.size() - 1; i>= first_keep ; i-- ) {
+    for ( size_t i = weights.size() - 1; i>= (size_t) first_keep ; i-- ) {
         if ( fabs(weights[i]) >= thresh ) {
             last_keep = (int)i;
             break;
